pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "chess_ai"
version = "0.1.0"
description = "Chess AI skeleton package"
dependencies = [
    "torch",
    "python-chess",
    "numpy",
    "lmdb",
]


[tool.setuptools.packages.find]
include = ["chess_ai*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
